% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{invwish}
\alias{invwish}
\alias{rinvwish}
\alias{dinvwish}
\title{Generate from and evaluate the density of the inverse Wishart distribution}
\usage{
rinvwish(v, S)

dinvwish(Sigma, v, S, logd = FALSE)
}
\arguments{
\item{v}{degrees of freedom (integer)}

\item{S}{\code{q * q} scale matrix}

\item{Sigma}{\code{q * q} matrix at which to evaluate the density}

\item{logd}{logical; if \code{TRUE} the logarithm of the density is returned}
}
\value{
For \code{rinvwish} a matrix \code{Sigma}, for \code{dinvwish} the (logarithm of) the density evaluation.
}
\description{
Generate from and evaluate the density of the inverse Wishart distribution
}
\examples{
set.seed(100)
q <- 5
v <- 10
S <- crossprod(matrix(rnorm(q*q), q, q))
Sigma <- rinvwish(v, S)
dinvwish(Sigma, v, S, logd = TRUE)
}
\references{
Karlsson, S. (2013). Forecasting with Bayesian vector autoregressions.
In G. Elliott & T. Timmermann (Eds.), Handbook of economic forecasting,
volume 2B. North Holland, Elsevier.
}
