% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matt}
\alias{matt}
\alias{rmatt}
\alias{dmatt}
\title{Generate from and evaluate the density of the matrix t distribution}
\usage{
rmatt(M, Q, P, v)

dmatt(X, M, Q, P, v, logd = FALSE)
}
\arguments{
\item{M}{\code{p * q} mean matrix}

\item{Q}{\code{q * q} covariance matrix}

\item{P}{\code{p * p} covariance matrix}

\item{v}{degrees of freedom (single-element numeric vector)}

\item{X}{\code{p * q} matrix at which to evaluate the density}

\item{logd}{logical; if \code{TRUE} the logarithm of the density is returned}
}
\value{
For \code{rmatt} a matrix \code{X}, for \code{dmatt} the (logarithm of) the density evaluation.
}
\description{
Generate from and evaluate the density of the matrix t distribution
}
\examples{
set.seed(100)
p <- 20
q <- 5
M <- matrix(rnorm(p*q), p, q)
P <- crossprod(matrix(rnorm(p*p), p, p))
Q <- crossprod(matrix(rnorm(q*q), q, q))
v <- 10
X <- rmatt(M, Q, P, v)
dmatt(X, M, Q, P, v, logd = TRUE)
}
\references{
Karlsson, S. (2013). Forecasting with Bayesian vector autoregressions.
In G. Elliott & T. Timmermann (Eds.), Handbook of economic forecasting,
volume 2B. North Holland, Elsevier.
}
