% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmatn}
\alias{rmatn}
\alias{dmatn}
\title{Generate from and evaluate the density of the matrix normal distribution}
\usage{
rmatn(M, Q, P)

dmatn(X, M, Q, P, logd = FALSE)
}
\arguments{
\item{M}{\code{p * q} mean matrix}

\item{Q}{\code{q * q} covariance matrix}

\item{P}{\code{p * p} covariance matrix}

\item{X}{\code{p * q} matrix at which to evaluate the density}

\item{logd}{logical; if \code{TRUE} the logarithm of the density is returned}
}
\value{
For \code{rmatn} a matrix \code{X}, for \code{dmatn} the (logarithm of) the density evaluation.
}
\description{
Generate from and evaluate the density of the matrix normal distribution
}
\examples{
set.seed(100)
p <- 20
q <- 5
M <- matrix(rnorm(p*q), p, q)
P <- crossprod(matrix(rnorm(p*p), p, p))
Q <- crossprod(matrix(rnorm(q*q), q, q))
X <- rmatn(M, Q, P)
dmatn(X, M, Q, P, logd = TRUE)
}
