% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmultn}
\alias{rmultn}
\alias{dmultn}
\title{Generate from and evaluate the density of the multivariate normal distribution}
\usage{
rmultn(m, Sigma)

dmultn(x, m, Sigma, logd = FALSE)
}
\arguments{
\item{m}{mean vector (length \code{p})}

\item{Sigma}{\code{p * p} covariance matrix}

\item{x}{\code{n * p} matrix at which to evaluate the density}

\item{logd}{logical; if \code{TRUE} the logarithm of the density is returned}
}
\value{
For \code{rmultn} a vector \code{x}, for \code{dmultn} the (logarithm of) the density evaluation.
set.seed(100)
p <- 20
m <- rnorm(p)
Sigma <- crossprod(matrix(rnorm(p*p), p, p))
x <- rmultn(m, Sigma)
dmultn(x, m, Sigma, logd = TRUE)
}
\description{
Generate from and evaluate the density of the multivariate normal distribution
}
